base of finding if a number is prime
we look first if we have a divider that is bigger than the sqr of a value. We need to change abs with sqr, I realized now.
if it does, then it must be a prime number.
then we look if the number is the leftover of a division of the n with a divider is 0, then it is dividable and not a prime number. if the number is bigger than 17^2 than we need to find the next prime number dividable